+++ packages/sources/coinbase-prime/src/config/index.ts
@@ -24,4 +24,10 @@ export const config = new AdapterConfig({
     required: true,
     sensitive: true,
   },
+  BACKGROUND_EXECUTE_MS: {
+    description:
+      'The amount of time the background execute should sleep before performing the next request',
+    type: 'number',
+    default: 10_000,
+  },
 })