+++ packages/composites/curve-3pool/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chainlink/curve-3pool-adapter
 
+## 3.0.70
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @chainlink/token-allocation-adapter@1.11.30
+
 ## 3.0.69
 
 ### Patch Changes