+++ packages/composites/reference-transform/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chainlink/reference-transform-adapter
 
+## 1.3.2
+
+### Patch Changes
+
+- Updated dependencies [[`ee90e49`](https://github.com/smartcontractkit/external-adapters-js/commit/ee90e49d6dedf9ff0d8598a71d298bb4dc70f263)]:
+  - @chainlink/ea-reference-data-reader@1.3.32
+
 ## 1.3.1
 
 ### Patch Changes